# Conflicts:
#	crypto/src/main/java/org/cryptimeleon/incentive/crypto/IncentiveSystem.java
#	crypto/src/main/java/org/cryptimeleon/incentive/crypto/Util.java
#	services/client/src/test/java/org/cryptimeleon/incentive/client/integrationtest/FullWorkflowTest.java
#	services/client/src/test/resources/application.properties
#	services/credit/src/main/java/org/cryptimeleon/incentive/services/credit/CreditController.java
#	services/credit/src/main/java/org/cryptimeleon/incentive/services/credit/CreditService.java
#	services/credit/src/test/java/org/cryptimeleon/incentive/services/credit/CreditIntegrationTest.java
#	services/issue/src/main/java/org/cryptimeleon/incentive/services/issue/IssueController.java
#	services/issue/src/main/java/org/cryptimeleon/incentive/services/issue/IssueService.java
#	services/settings.gradle

+++ services/client/src/main/java/org/cryptimeleon/incentive/client/dto/BasketItemDto.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class BasketItemDto {
     String id;
     String title;
-    long price;
+    int price;
 }
